requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
blas=1.0=mkl
brotli-python=1.0.9=py39h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.11.17=hbcca054_0
certifi=2023.11.17=pyhd8ed1ab_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.1.1=pypi_0
cryptography=41.0.3=py39hdda0065_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
docstring-parser=0.16=pypi_0
einops=0.7.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.14.0=pypi_0
fonttools=4.43.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.3.1=pypi_0
fvcore=0.1.5.post20221221=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpcio=1.63.0=pypi_0
huggingface-hub=0.23.0=pypi_0
idna=3.4=py39h06a4308_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.1.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
iopath=0.1.10=pypi_0
ipykernel=6.26.0=pyhf8b6a83_0
ipython=8.18.1=pyh31011fe_1
jedi=0.19.1=pyhd8ed1ab_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.5.0=py39hf3d152e_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
llvmlite=0.41.1=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.0=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mpmath=1.3.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
numba=0.58.1=pypi_0
numpy=1.26.1=pypi_0
numpy-base=1.26.0=py39hb5e798b_0
opencv-python=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.12=h7f8727e_0
packaging=23.2=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.1.0=pypi_0
pip=23.3=py39h06a4308_0
platformdirs=4.0.0=pyhd8ed1ab_0
portalocker=2.8.2=pypi_0
prompt-toolkit=3.0.41=pyha770c72_0
protobuf=4.24.4=pypi_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.17.2=pyhd8ed1ab_0
pykan=0.0.2=pypi_0
pynndescent=0.5.11=pypi_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.12.1=py3.9_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py39h6a678d5_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.31.0=py39h06a4308_0
safetensors=0.4.3=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=pypi_0
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.19.1=pypi_0
torchaudio=0.12.1=py39_cu113
torchvision=0.13.1=py39_cu113
tornado=6.1=py39hb9d737c_3
tqdm=4.66.1=pypi_0
traitlets=5.14.0=pyhd8ed1ab_0
transformers=4.40.2=pypi_0
typed-argument-parser=1.10.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023c=h04d1e81_0
umap-learn=0.5.5=pypi_0
urllib3=1.26.18=py39h06a4308_0
wcwidth=0.2.12=pyhd8ed1ab_0
werkzeug=3.0.3=pypi_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.2=h5eee18b_0
yacs=0.1.8=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
